pyproject.toml
[project]
name = "anttec-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.5",
    "groq<1.0.0",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
